requirements.txt
blinker==1.6.2
click==8.1.4
Flask==2.3.2
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
python-dotenv==1.0.0
Werkzeug==2.3.6
waitress~=2.1.2
openai~=0.27.8
pymongo~=4.4.0
certifi~=2023.5.7